+++ __tests__/get-wprs-civl.test.ts
@@ -1,0 +1,13 @@
+import { getWprs } from "@/utils/calculate-wprs";
+
+describe("Get WPRS for CIVL comps", () => {
+  it("should reject a comp that lies in the past", async () => {
+    const url = "https://civlcomps.org/event/german-open-2023";
+    const res = await getWprs(url);
+
+    expect(res).toHaveProperty("error");
+    if ("error" in res) {
+      expect(res.error).toBe("PAST_EVENT");
+    }
+  }, 10000);
+});