+++ src/components/ListRankings.tsx
@@ -1,7 +1,6 @@
-import { type GetWPRS } from "@/utils/calculate-wprs";
+import { type Forecast } from "@/utils/calculate-wprs";
 
-export function ListRankings({ data }: { data: GetWPRS }) {
-  if (data === 0) return <></>;
+export function ListRankings({ data }: { data: Forecast }) {
   const listRanking = () => {
     if (!data.confirmed?.WPRS.length) return;
     return data?.confirmed?.WPRS.map((el, i) => {
@@ -18,7 +17,7 @@ export function ListRankings({ data }: { data: GetWPRS }) {
 
   return (
     <>
-      <div className="collapse-arrow rounded-box collapse mt-3 border ">
+      <div className="collapse collapse-arrow rounded-box mt-3 border ">
         <input type="checkbox" />
         <div className="collapse-title font-medium">
           Show points for every position